requirements.txt
pandas
python-dotenv
cassandra-driver
openpyxl
astrapy
